+++ Laboratory_work_7/lab_7/executable_2.cpp
@@ -9,12 +9,16 @@
 
 using namespace std;
 
+// false -- writing in pipe is NOT finished
+// true -- writing in pipe is finished
+bool pipe_write_is_finished = false; // end of writing indication
 void LocalHandler (int local_int);
 
 int main(int argc, char *argv[])
 {
 	FILE* output_file_2 = fopen(argv[3], "w"); // file opening w/ write flag
 	int sig; // for sigwait
+	int pipe_read_is_done = 1; // if > 0, pipe read is NOT done, if < 0, is done
 	int fildes[2]; // pipe channels handles, fildes[0] -- read from pipe, fildes[1] -- write to pipe
 	char ch; // char buffer
 	sigset_t b_set;
@@ -34,11 +38,16 @@ int main(int argc, char *argv[])
 	
 	sigwait(&set, &sig); // child process 2 waits child process 1 (1st waiting before reading)
 	
-	while(read(fildes[0], &ch, 1) > 0) // reading from pipe
+	//pipe_read_is_done = read(fildes[0], &ch, 1); // reading from pipe
+	while ((pipe_read_is_done = read(fildes[0], &ch, 1)) > 0 || pipe_write_is_finished == false) // while pipe writing by parent process is not finished & pipe not done
 	{
-		fputc(ch, output_file_2); // writing to the file
-		kill(0, SIGUSR1); // child process 2 give signal to child process 1
-		sigwait(&set, &sig); // child process 2 waits child process 1
+		if (pipe_read_is_done > 0) // additional check pipe if is done (if previous while there was ||, not &&)
+		{
+			fputc(ch, output_file_2); // writing to the file
+			kill(0, SIGUSR1); // child process 2 give signal to child process 1
+			sigwait(&set, &sig); // child process 2 waits child process 1
+		}
+		//pipe_read_is_done = read(fildes[0], &ch, 1); // reading from pipe
 	}
 	
 	kill(0, SIGUSR1); // child process 2 give signal to child process 1 FINAL BEFORE TERMINATION
@@ -48,4 +57,6 @@ int main(int argc, char *argv[])
 }
 
 void LocalHandler (int local_int)
-{}
+{
+	pipe_write_is_finished = true; // pipe writing is finished
+}
